ignore net-dns/ddclient-3.10.0_2
We need to ignore net-dns/ddclient-3.10.0, since repology reports
'net-dns/ddclient-3.10.0_2' as newest version, which is an error due
upstream decided to call rc releases _1, _2 and so on.

Technically, net-dns/ddclient-3.10.0_2 is newer, but from upstream point
of view it's only rc. So we need to ignore it for now. Repology also
don't to add an excpetion here.

Signed-off-by: Conrad Kostecki <conrad@kostecki.com>