ignored-packages: ignore app-emulation/ruffle
Seen a user think that 20240119 (reported by repology) was newer than
0_p20240319 given repology does not understand that type of versioning.

Even if changed that so it works, reports are also not interesting
given this is nightlies and a version added today could be marked
outdated tomorrow if a distro just decided to bump on that day and
getting into a bump war is pointless.

This should be removed when ruffle start getting real releases
(currently ruffle --version reports 0.1.0, so the 0_ is a placeholder
for potential future normal versioning).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>